gracefully handle strings passed to the NumberFormatter
I was getting `undefined is not a function` errors due to my model number attr being represented as a string. e.g. 

```javascript
model.get('count') //-> "33"
```

A simple fix would just be to parseFloat the passed 'number' to ensure that it is a number.

I don't think there is an issue with this change?